pyproject.toml
[project]
name = "firefly-bills-analyzer"
version = "0.1.0"
description = "Analyzes your Firefly III transaction history to automatically identify recurring payments and create subscriptions (bills) via the Firefly III API. Designed for cash flow planning across the full year, including low-frequency bills such as quarterly and annual payments."
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "bandit",
    "pymarkdownlnt>=0.9.36",
    "complexipy",
    "pytest",
    "pytest-cov",
    "pre-commit",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
strict = true

[tool.bandit]
skips = []

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
